backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.3.4
asyncpg==0.29.0
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background tasks
celery==5.3.4
redis==5.0.1

# Bot dependencies (your existing ones)
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.3.2
anthropic==0.7.8
pypdf==3.17.4

# Notifications
twilio==8.10.0
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
